refactor runs mgr to be inline with other recent refactors around features
