+++ ui/src/routes/workspaces.$workspaceId_.runs.$tab.tsx
@@ -1,7 +1,7 @@
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from "@tanstack/react-router";
 
-import { Runs } from '@flow/features/Runs'
+import { RunsManager } from "@flow/features/WorkspaceRuns";
 
-export const Route = createFileRoute('/workspaces/$workspaceId_/runs/$tab')({
-  component: () => <Runs />,
-})
+export const Route = createFileRoute("/workspaces/$workspaceId_/runs/$tab")({
+  component: () => <RunsManager />,
+});
